Fix binary-arch target to not require Build-Depends-Indep installed
Disable docbook support in configure so that it does not fail when
we only want build-arch or binary-arch.
On build-indep build doc/docbook and pass DOCBOOK2HTML to it (because
it's no longer set by configure).
On install-indep install doc/docbook, remove install target dependency,
call dh_prep, and populate the debian/tmp directory.

Closes: #562964